test(codex): split codegraph trust boundary coverage
