+++ scripts/check-third-party-notices.mjs
@@ -4,6 +4,8 @@ import { dirname, join } from "node:path"
 import { spawnSync } from "node:child_process"
 import { fileURLToPath, pathToFileURL } from "node:url"
 
+import { CODEX_COMPONENT_NOTICE_REQUIREMENTS } from "./third-party-notice-requirements.mjs"
+
 const repoRoot = join(dirname(fileURLToPath(import.meta.url)), "..")
 const WINDOWS_CMD_SHIM_COMMANDS = new Set(["npm", "npx"])
 
@@ -61,42 +63,6 @@ const CODEX_AGGREGATE_COMPONENTS = [
   "pi-rules",
   "picomatch",
   "posthog-node",
-]
-
-const CODEX_COMPONENT_NOTICE_REQUIREMENTS = [
-  {
-    path: "packages/omo-codex/plugin/components/codegraph",
-    requiredTerms: ["@colbymchenry/codegraph", "Node.js runtime", "MIT license"],
-    forbiddenTerms: ["packages/omo-codex/THIRD-PARTY-NOTICES.md"],
-  },
-  {
-    path: "packages/omo-codex/plugin/components/comment-checker",
-    requiredTerms: ["pi-comment-checker", "@code-yeongyu/comment-checker"],
-  },
-  {
-    path: "packages/omo-codex/plugin/components/lsp",
-    requiredTerms: ["pi-lsp-client"],
-  },
-  {
-    path: "packages/omo-codex/plugin/components/rules",
-    requiredTerms: ["pi-rules", "picomatch"],
-  },
-  {
-    path: "packages/omo-codex/plugin/components/start-work-continuation",
-    requiredTerms: [],
-  },
-  {
-    path: "packages/omo-codex/plugin/components/telemetry",
-    requiredTerms: ["posthog-node", "@oh-my-opencode/telemetry-core"],
-  },
-  {
-    path: "packages/omo-codex/plugin/components/ultrawork",
-    requiredTerms: [],
-  },
-  {
-    path: "packages/omo-codex/plugin/components/ulw-loop",
-    requiredTerms: [],
-  },
 ]
 
 const ROOT_SHIP_REQUIRED_PATHS = [
